# perf: Extract properties as `Nullable(String)` instead of raw values