## Problem

Using `JSONExtractRaw` to extract values has a couple of drawbacks:

1. it requires additional post-processing to discriminate `NULL` values from valid strings,
2. it requires additional post-processing to strip leading and trailing quotes from strings,
3. string values are returned as escaped values, which require additional post-processing to handle correctly, compare:

```sql
SELECT JSONExtractRaw('{"x": "a \\"quoted\\" value"}', 'x')

Query id: b228e394-cbb6-4692-bcef-bc7c43afca36

   ┌─JSONExtractRaw('{"x": "a \\"quoted\\" value"}', 'x')─┐
1. │ "a \"quoted\" value"                                 │
   └──────────────────────────────────────────────────────┘

1 row in set. Elapsed: 0.001 sec. 
```

vs.

```sql
SELECT JSONExtractString('{"x": "a \\"quoted\\" value"}', 'x')

Query id: 84eaf80e-746f-4f97-a6dd-c53515e860a3

   ┌─JSONExtractString('{"x": "a \\"quoted\\" value"}', 'x')─┐
1. │ a "quoted" value                                        │
   └─────────────────────────────────────────────────────────┘

1 row in set. Elapsed: 0.001 sec. 
```

Handling these special cases incurs performance overhead, and/or not handling them leads to unintended effects on the extracted values.

## Changes

This changes the `JSONExtract` call to extract a `Nullable(String)` rather than the raw value. This is consistent with the way the values are being materialized into property groups (#24152) and has a noticeable performance impact as shown in these benchmarks (some values redacted because customer data):

```sql
SELECT count()
FROM events
WHERE
    (toDate(timestamp) >= '2024-05-01')
    AND (toDate(timestamp) <= '2024-05-31')
    AND (team_id = …)
    AND (replaceRegexpAll(nullIf(nullIf(JSONExtractRaw(properties, 'category'), ''), 'null'), '^"|"$', '') = …)

Query id: d595c049-0407-426b-aeeb-a0fa515287f7

   ┌───count()─┐
1. │ 113617422 │ -- 113.62 million
   └───────────┘

1 row in set. Elapsed: 34.112 sec. Processed 284.40 million rows, 571.08 GB (8.34 million rows/s., 16.74 GB/s.)
Peak memory usage: 534.84 MiB.
```

vs.

```sql
SELECT count()
FROM events
WHERE
    (toDate(timestamp) >= '2024-05-01')
    AND (toDate(timestamp) <= '2024-05-31')
    AND (team_id = …)
    AND (JSONExtract(properties, 'category', 'Nullable(String)') = …)

Query id: fa065dd3-0483-4cbb-ae98-d9beed964571

   ┌───count()─┐
1. │ 113617422 │ -- 113.62 million
   └───────────┘

1 row in set. Elapsed: 29.918 sec. Processed 284.40 million rows, 571.08 GB (9.51 million rows/s., 19.09 GB/s.)
Peak memory usage: 530.59 MiB.
```

Comparing the profile events of the two queries shows a **~12% reduction** in `OSCPUVirtualTimeMicroseconds` and `UserTimeMicroseconds`, while other significant events remained relatively constant, which is consistent with the overall query execution speedup.

<details><summary>Profile Events Comparison Query and Results</summary>

```sql
SELECT
    query_events.*,
    events.description
FROM
(
    WITH
        query_id = 'd595c049-0407-426b-aeeb-a0fa515287f7' AS first,
        query_id = 'fa065dd3-0483-4cbb-ae98-d9beed964571' AS second
    SELECT
        event,
        anyIf(value, first) AS v1,
        anyIf(value, second) AS v2,
        v2 - v1 AS difference,
        v2 / v1 AS rel_diff
    FROM clusterAllReplicas(default, system, query_log) AS queries
    ARRAY JOIN
        ProfileEvents.keys AS event,
        ProfileEvents.values AS value
    WHERE (first OR second) AND (type != 'QueryStart')
    GROUP BY ALL
    ORDER BY ALL ASC
) AS query_events
LEFT JOIN system.events AS events ON query_events.event = events.event
FORMAT Markdown
```

| event | v1 | v2 | difference | rel_diff | description |
|:-|-:|-:|-:|-:|:-|
| ArenaAllocBytes | 118784 | 118784 | 0 | 1 | Number of bytes allocated for memory Arena (used for GROUP BY and similar operations) |
| ArenaAllocChunks | 29 | 29 | 0 | 1 | Number of chunks allocated for memory Arena (used for GROUP BY and similar operations) |
| AsynchronousRemoteReadWaitMicroseconds | 56169 | 56247 | 78 | 1.0013886663462053 | Time spent in waiting for asynchronous remote reads. |
| BackgroundLoadingMarksTasks | 486 | 474 | -12 | 0.9753086419753086 | Number of background tasks for loading marks |
| CachedReadBufferCreateBufferMicroseconds | 380142 | 394628 | 14486 | 1.038106812717353 | Prepare buffer time |
| CachedReadBufferReadFromCacheBytes | 27248583579 | 27247819102 | -764477 | 0.9999719443398669 | Bytes read from filesystem cache |
| CachedReadBufferReadFromCacheHits | 218900 | 218893 | -7 | 0.999968021927821 | Number of times the read from filesystem cache hit the cache. |
| CachedReadBufferReadFromCacheMicroseconds | 7397192 | 7743056 | 345864 | 1.0467561204305633 | Time reading from filesystem cache |
| CompressedReadBufferBlocks | 636587 | 636576 | -11 | 0.9999827203508711 | Number of compressed blocks (the blocks of data that are compressed independent of each other) read from compressed sources (files, network). |
| CompressedReadBufferBytes | 569109013530 | 569109980426 | 966896 | 1.0000016989644813 | Number of uncompressed bytes (the number of bytes after decompression) read from compressed sources (files, network). |
| ContextLock | 1698 | 1671 | -27 | 0.9840989399293286 | Number of times the lock of Context was acquired or tried to acquire. This is global lock. |
| ContextLockWaitMicroseconds | 1 | 1 | 0 | 1 | Context lock wait time in microseconds |
| CreatedReadBufferOrdinary | 18707 | 18707 | 0 | 1 | Number of times ordinary read buffer was created for reading data (while choosing among other read methods). |
| DiskReadElapsedMicroseconds | 7256365 | 7600087 | 343722 | 1.0473683448944477 | Total time spent waiting for read syscall. This include reads from page cache. |
| FileOpen | 17723 | 17810 | 87 | 1.0049088754725497 | Number of files opened. |
| FileSegmentCompleteMicroseconds | 95 | 99 | 4 | 1.0421052631578946 | Duration of FileSegment::complete() in filesystem cache |
| FileSegmentFailToIncreasePriority | 35 | 48 | 13 | 1.3714285714285714 | Number of times the priority was not increased due to a high contention on the cache lock |
| FileSegmentHolderCompleteMicroseconds | 4952 | 6496 | 1544 | 1.3117932148626819 | File segments holder complete() time |
| FileSegmentLockMicroseconds | 692 | 887 | 195 | 1.2817919075144508 | Lock file segment time |
| FileSegmentUseMicroseconds | 15129 | 15447 | 318 | 1.0210192345825897 | File segment use() time |
| FileSegmentWaitReadBufferMicroseconds | 393521 | 408188 | 14667 | 1.0372712002663136 | Metric per file segment. Time spend waiting for internal read buffer (includes cache waiting) |
| FilesystemCacheGetOrSetMicroseconds | 98279 | 108830 | 10551 | 1.1073576247214563 | Filesystem cache getOrSet() time |
| FilesystemCacheHoldFileSegments | 18761 | 18716 | -45 | 0.9976014071744577 | Filesystem cache file segments count, which were hold |
| FilesystemCacheLockKeyMicroseconds | 43 | 20 | -23 | 0.46511627906976744 | Lock cache key time |
| FilesystemCacheLockMetadataMicroseconds | 91 | 60 | -31 | 0.6593406593406593 | Lock filesystem cache metadata time |
| FilesystemCacheUnusedHoldFileSegments | 14237 | 14192 | -45 | 0.9968392217461544 | Filesystem cache file segments count, which were hold, but not used (because of seek or LIMIT n, etc) |
| FunctionExecute | 1548250 | 1095043 | -453207 | 0.7072778943968997 | Number of SQL ordinary function calls (SQL functions are called on per-block basis, so this number represents the number of blocks). |
| IOBufferAllocBytes | 1062063622 | 1039160183 | -22903439 | 0.9784349651700998 | Number of bytes allocated for IO buffers (for ReadBuffer/WriteBuffer). |
| IOBufferAllocs | 1262 | 1235 | -27 | 0.9786053882725833 | Number of allocations of IO buffers (for ReadBuffer/WriteBuffer). |
| InitialQuery | 1 | 1 | 0 | 1 | Same as Query, but only counts initial queries (see is_initial_query). |
| LogDebug | 94 | 94 | 0 | 1 | Number of log messages with level Debug |
| LogTrace | 2954 | 2954 | 0 | 1 | Number of log messages with level Trace |
| MarkCacheHits | 486 | 474 | -12 | 0.9753086419753086 | Number of times an entry has been found in the mark cache, so we didn\'t have to load a mark file. |
| MergeTreePrefetchedReadPoolInit | 12 | 13 | 1 | 1.0833333333333333 | Time spent preparing tasks in MergeTreePrefetchedReadPool |
| MetadataFromKeeperCacheHit | 2320 | 2275 | -45 | 0.9806034482758621 | Number of times an object storage metadata request was answered from cache without making request to Keeper |
| NetworkReceiveElapsedMicroseconds | 1463 | 1443 | -20 | 0.9863294600136705 | Total time spent waiting for data to receive or receiving data from network. Only ClickHouse-related network interaction is included, not by 3rd party libraries. |
| NetworkSendBytes | 2755902 | 2441311 | -314591 | 0.885848263109501 | Total number of bytes send to network. Only ClickHouse-related network interaction is included, not by 3rd party libraries. |
| NetworkSendElapsedMicroseconds | 28810 | 28823 | 13 | 1.000451232211038 | Total time spent waiting for data to send to network or sending data to network. Only ClickHouse-related network interaction is included, not by 3rd party libraries. |
| OSCPUVirtualTimeMicroseconds | 976863710 | 855566858 | -121296852 | 0.8758303223281783 | CPU time spent seen by OS. Does not include involuntary waits due to virtualization. |
| OSCPUWaitMicroseconds | 7044804 | 6757738 | -287066 | 0.9592513858440916 | Total time a thread was ready for execution but waiting to be scheduled by OS, from the OS point of view. |
| OSIOWaitMicroseconds | 40000 | 10000 | -30000 | 0.25 | Total time a thread spent waiting for a result of IO operation, from the OS point of view. This is real IO that doesn\'t include page cache. |
| OSReadBytes | 675840 | 491520 | -184320 | 0.7272727272727273 | Number of bytes read from disks or block devices. Doesn\'t include bytes read from page cache. May include excessive data due to block size, readahead, etc. |
| OSReadChars | 27281061270 | 27277222798 | -3838472 | 0.9998592990220574 | Number of bytes read from filesystem, including page cache. |
| OSWriteBytes | 749568 | 757760 | 8192 | 1.010928961748634 | Number of bytes written to disks or block devices. Doesn\'t include bytes that are in page cache dirty pages. May not include data that was written by OS asynchronously. |
| OSWriteChars | 1583214 | 1581972 | -1242 | 0.999215519822336 | Number of bytes written to filesystem, including page cache. |
| OpenedFileCacheHits | 984 | 897 | -87 | 0.9115853658536586 | Number of times a file has been found in the opened file cache, so we didn\'t have to open it again. |
| OpenedFileCacheMicroseconds | 13940 | 15867 | 1927 | 1.138235294117647 | Amount of time spent executing OpenedFileCache methods. |
| OpenedFileCacheMisses | 17723 | 17810 | 87 | 1.0049088754725497 | Number of times a file has been found in the opened file cache, so we had to open it again. |
| PartsLockHoldMicroseconds | 341 | 344 | 3 | 1.0087976539589443 | Total time spent holding data parts lock in MergeTree tables |
| QueriesWithSubqueries | 1 | 1 | 0 | 1 | Count queries with all subqueries |
| Query | 1 | 1 | 0 | 1 | Number of queries to be interpreted and potentially executed. Does not include queries that failed to parse or were rejected due to AST size limits, quota limits or limits on the number of simultaneously running queries. May include internal queries initiated by ClickHouse itself. Does not count subqueries. |
| RWLockAcquiredReadLocks | 1 | 1 | 0 | 1 | Number of times a read lock was acquired (in a heavy RWLock). |
| ReadBufferFromFileDescriptorRead | 218900 | 218893 | -7 | 0.999968021927821 | Number of reads (read/pread) from a file descriptor. Does not include sockets. |
| ReadBufferFromFileDescriptorReadBytes | 27248583579 | 27247819102 | -764477 | 0.9999719443398669 | Number of bytes read from file descriptors. If the file is compressed, this will show the compressed data size. |
| ReadCompressedBytes | 27247574491 | 27247234102 | -340389 | 0.9999875075486035 | Number of bytes (the number of bytes before decompression) read from compressed sources (files, network). |
| RealTimeMicroseconds | 1056093728 | 926102935 | -129990793 | 0.8769135829959214 | Total (wall clock) time spent in processing (queries and other tasks) threads (note that this is a sum). |
| RemoteFSBuffers | 464 | 455 | -9 | 0.9806034482758621 | Number of buffers created for asynchronous reading from remote filesystem |
| RemoteFSCancelledPrefetches | 5 | 4 | -1 | 0.8 | Number of cancelled prefecthes (because of seek) |
| RemoteFSPrefetchedBytes | 18545564081 | 18544760677 | -803404 | 0.9999566794519438 | Number of bytes from prefecthed buffer |
| RemoteFSPrefetchedReads | 150594 | 150593 | -1 | 0.9999933596292017 | Number of reads from prefecthed buffer |
| RemoteFSPrefetches | 150660 | 150654 | -6 | 0.9999601752289924 | Number of prefetches made with asynchronous reading from remote filesystem |
| RemoteFSSeeks | 7719 | 7708 | -11 | 0.9985749449410546 | Total number of seeks for async buffer |
| RemoteFSUnprefetchedBytes | 8703019498 | 8703058425 | 38927 | 1.0000044728154418 | Number of bytes from unprefetched buffer |
| RemoteFSUnprefetchedReads | 68240 | 68239 | -1 | 0.9999853458382181 | Number of reads from unprefetched buffer |
| RemoteFSUnusedPrefetches | 61 | 57 | -4 | 0.9344262295081968 | Number of prefetches pending at buffer destruction |
| Seek | 14181 | 14181 | 0 | 1 | Number of times the \'lseek\' function was called. |
| SelectQueriesWithSubqueries | 1 | 1 | 0 | 1 | Count SELECT queries with all subqueries |
| SelectQuery | 1 | 1 | 0 | 1 | Same as Query, but only for SELECT queries. |
| SelectedBytes | 571076466023 | 571076466023 | 0 | 1 | Number of bytes (uncompressed; for columns as they stored in memory) SELECTed from all tables. |
| SelectedMarks | 140202 | 140202 | 0 | 1 | Number of marks (index granules) selected to read from a MergeTree table. |
| SelectedParts | 44 | 44 | 0 | 1 | Number of data parts selected to read from a MergeTree table. |
| SelectedRanges | 44 | 44 | 0 | 1 | Number of (non-adjacent) ranges in all data parts selected to read from a MergeTree table. |
| SelectedRows | 284397897 | 284397897 | 0 | 1 | Number of rows SELECTed from all tables. |
| SoftPageFaults | 1539022 | 1068522 | -470500 | 0.6942863714748717 | The number of soft page faults in query execution threads. Soft page fault usually means a miss in the memory allocator cache, which requires a new memory mapping from the OS and subsequent allocation of a page of physical memory. |
| SynchronousRemoteReadWaitMicroseconds | 2538611 | 2661650 | 123039 | 1.0484670554094346 | Time spent in waiting for synchronous remote reads. |
| SystemTimeMicroseconds | 11817303 | 11138050 | -679253 | 0.9425204718877057 | Total time spent in processing (queries and other tasks) threads executing CPU instructions in OS kernel mode. This is time spent in syscalls, excluding waiting time during blocking syscalls. |
| ThreadpoolReaderPrepareMicroseconds | 513 | 4570 | 4057 | 8.908382066276802 | Time spent on preparation (e.g. call to reader seek() method) |
| ThreadpoolReaderReadBytes | 27248583579 | 27247819102 | -764477 | 0.9999719443398669 | Bytes read from a threadpool task in asynchronous reading |
| ThreadpoolReaderSubmitLookupInCacheMicroseconds | 328865 | 346951 | 18086 | 1.0549952108007845 | How much time we spent checking if content is cached |
| ThreadpoolReaderSubmitReadSynchronously | 150660 | 150654 | -6 | 0.9999601752289924 | How many times we haven\'t scheduled a task on the thread pool and read synchronously instead |
| ThreadpoolReaderSubmitReadSynchronouslyBytes | 18545564081 | 18544760677 | -803404 | 0.9999566794519438 | How many bytes were read synchronously |
| ThreadpoolReaderSubmitReadSynchronouslyMicroseconds | 6308188 | 6567264 | 259076 | 1.0410697969052285 | How much time we spent reading synchronously |
| ThreadpoolReaderTaskMicroseconds | 8709905 | 9081807 | 371902 | 1.0426987435569044 | Time spent getting the data in asynchronous reading |
| UserTimeMicroseconds | 965080961 | 844460503 | -120620458 | 0.8750151926372942 | Total time spent in processing (queries and other tasks) threads executing CPU instructions in user mode. This includes time CPU pipeline was stalled due to main memory access, cache misses, branch mispredictions, hyper-threading, etc. |
| WaitMarksLoadMicroseconds | 149 | 129 | -20 | 0.8657718120805369 | Time spent loading marks |
| WaitPrefetchTaskMicroseconds | 158 | 175 | 17 | 1.1075949367088607 | Time spend waiting for prefetched reader |
</details>

## Does this work well for both Cloud and self-hosted?

Yes.

## How did you test this code?

<!-- Briefly describe the steps you took. -->
<!-- Include automated tests if possible, otherwise describe the manual testing routine. -->